Add mariadb support
